+++ CHANGELOG.md
@@ -11,7 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Add `pallet-assets` on Polkadot People to support HOLLAR (and more assets in the future) ([polkadot-fellows/runtimes/pull/962](https://github.com/polkadot-fellows/runtimes/pull/962))
-- Collectives/Asset Hub Polkadot: Allow Architects track (rank 4+ Fellowship members) to manage Fellowship Treasury and Salary on Asset Hub via XCM ([polkadot-fellows/runtimes/pull/1072](https://github.com/polkadot-fellows/runtimes/pull/1072))
+- Collectives/Asset Hub Polkadot: Allow Architects track (rank 4+ Fellowship members) to manage Fellowship Treasury and Salary on Asset Hub via XCM. The Architects origin is represented as `[Plurality { id: Technical, part: Voice }, GeneralIndex(4)]` ([polkadot-fellows/runtimes/pull/1072](https://github.com/polkadot-fellows/runtimes/pull/1072))
 - Encointer: add pallets `offline-payment` and `reputation-rings` and fixes ([polkadot-fellows/runtimes/pull/1082](https://github.com/polkadot-fellows/runtimes/pull/1082))
 - Add `pallet_multi_asset_bounties` pallet to Asset Hub Kusama and Asset Hub Polkadot for multi-asset bounty support ([#1070](https://github.com/polkadot-fellows/runtimes/pull/1070))